LJ-1 removed redundant functional, refactored gui packages
