--- src/main/java/org/vector/littlejourney/Transport.java
@@ -1,9 +1,0 @@
-package org.vector.littlejourney;
-
-public class Transport {
-
-   private String type;
-
-   private String name;
-
-}
